Watch useWeekly vs. data
This allows us to re-render appropriately when the user changes the volume chart from day/week but does not re-render when data is generally changed. Data is only fetched once, and does not update live. Previously we watched data knowing that it would be changed when the weekly/daily option was changed and can be more clearly watched this way.